Fixing a few more places where RGIS and the DBlib was not quite 64bit ready.

+++ DBlib/src/DBGCont2Net.C
@@ -85,7 +85,7 @@ DBInt DBGridCont2Network (DBObjData *gridData,DBObjData *netData, bool downhill)
 	rowNumFLD->Int (layerRec,gridIF->RowNum ());
 	colNumFLD->Int (layerRec,gridIF->ColNum ());
 
-	dataRec = new DBObjRecord ("NetLookupGridRecord",gridIF->RowNum () * gridIF->ColNum () * sizeof (DBInt),sizeof (DBInt));
+	dataRec = new DBObjRecord ("NetLookupGridRecord",((size_t) gridIF->RowNum ()) * gridIF->ColNum () * sizeof (DBInt),sizeof (DBInt));
 	if (dataRec == (DBObjRecord *) NULL)
 		{
 		if (zGridIF != (DBGridIF *) NULL) delete zGridIF;
